backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
requests
python-docx
PyPDF2
python-multipart
supabase
pycryptodome
jinja2
httpx
markupSafe
